Keyboard improvements
Added ability to disable keys and change the way that they are rendered.

Modified the thumb keys to add tilt and large keys.
Added 7th col to the main section.
Moved the thumb section for the addition of the 7th col.
Added 2 functions that generate the xy position for the keys.